Simplify style loaders for server and prod webpack bundles
Since we are using styled-components exclusively at this
point, this commit strips away the custom loaders from
the server webpack config. We also do not need to use
ExtractTextPlugin anymore as we will extract critical
CSS as part of server side rendering and will keep the
rest of the styling in our bundle for now.