Fix development server
- The first thing needing to be done was modifying the
shared template that we are using for both server side rendering
as well as the development environment. I created a custom webpack
plugin that is added to webpack.dev.babel.js that strips away
the interpolation variables that is required for SSR so that they
do not appear in the browser during development.
- Once index.html fit our needs in development, we could return the
dev middleware back to it's original settings.

NOTE: It is important to understand that we are using webpack-dev-middleware,
not webpack-dev-server. webpack-dev-server allows us to set up the same type of behavior as webpack-dev-server using our own custom express server. I am still uncertain if we could not achieve the same dev environment with the pre-baked server or not.