Create /search/movie endpoint for finding new movies
I decided to route search requests through our application rather than
directly from the client to tmdb to create API consistency for our front-end
code. Otherwise, logic that would be preferred only on the server would
leak out to the client. This seemed undesirable. Doing it this way allows
the client to interact with the data the exact same way it does with our API.