Restructure models and model services
Because of the way that models are initialized by Sequelize, it makes it difficult to create new methods that retrieve related data directly in the model file. moviesService was simply a place to put these custom model methods. As a result, it made sense to nest this module and others like it into a models/services directory. The top level services directory will be reserved for other functionality that assists controllers but that is not directly interacting with the database layer.