Refactor Search component to accommodate autofocus
Applying focus on an element dynamically requires using innerRef (ref for non-styled components) to retrieve the input and set it as a property on the component. Then using componentWillUpdate or componentWillMount, you can call `focus` on the input. Because we are using lifecycle methods and keeping state on the component, it made more sense as a class than a stateless function.