Force App to update on location change
The react application was not updating on any router location change, standard or redirect. I discovered it was because of the `connect` method in react-redux implementing `shouldComponentUpdate` and not seeing any difference in the App component's props on route navigation (https://github.com/ReactTraining/react-router/blob/master/packages/react-router/docs/guides/blocked-updates.md#shouldcomponentupdate).

According to the documentation, 'The key to avoiding blocked re-renders after location changes is to pass the blocking component the location object as a prop. This will be different whenever the location changes, so comparisons will detect that the current and next location are different.'

There are two ways to inject location as a prop to a connected component: the option I chose was to use a pathless Route component to render the App component.