Enable favoriting library movie
This feature addition came with a fairly hefty refactor of the movie model service. We needed to create a more consistent way of retrieving movies so that we had the same data to use on the frontend, specifically the id from the UserMovies table. As a result, I was able to reuse a good portion of the service method code between each other.