Fix issues caused by SubMenu addition
The refactored DashboardPage refreshes the library listing every time that new props are passed to it. This is necessary so that we can determine which library view to show the user. However, this caused favoriting movies to break because the updated state after favoriting was reverted to the cached library state on the next render cycle. The fix was rather simple as we simply need to mark the dashboard resource as dirty after any changes that affect the list of movies. We also find the other SubMenu views that are affected and invalidate those resourceCache keys as well so that they will retrieve fresh data from the server the next time they are selected.