Add ability to remove movie from library (not only search)
- Added same behavior to library search results
- Corrected behavior of hideSearchHandler in dashboard. Now closing search will return correct library data depending on current submenu selection
- Patched behavior of favoriting over to library search results

One lesson learned from this work is that any action created for a library item in the dashboard will need to be replicated in the library SearchResult component. This is necessary because search results and library are two distinct data sets held in state and therefore need to be updated independently of each other.