Update custom ConnectedRouter
We are not currently using this component as it didn't solve the problem I was hoping it would (providing query params parsed wherever location object was accessed). This was because React Router passes the location prop to child components, and I need this value for understanding what submenu page to show on the dashboard.